Add UI to RPS game
Added buttons to index.html, as well as two divs to output the score and results. Reworked js to add event callers to the buttons and change the outputs from logs to text content in the divs. Added if statement to playRound function in order to stop at 5 wins and declare the winner. Game works off of playRound function rather than playFive now.